fizzbuzz.php   rewrite the page to use PHP on the server                   - add PHP code for cross-platform compatibility index,php      rewrite the page to use PHP on the server                   - add PHP code for cross-platform compatibility testmysql.php  rewrite the page to use PHP on the server                   - add PHP code for cross-platform compatibility ***
One problem with Web development is that the various browsers are
not 100% compatible. There are various small things which have to
be corrected for the page to render correctly in each browser.
Those things might be corrected if we use an application framework,
but what do we do if we aren't using one? How do we get the page to
render correctly if we are using generic PHP or generic JavaScript?

PHP executes on the server. JavaScript executes on the client.
JavaScript on the client is able to access the information to allow
us to provide cross-platform compatibility, but is unable to share
information directly with PHP running on the server.

In order to successfully render the Web pages, we need to identify
the client device so that we know which items on-screen need to be
corrected. We would need to know the browser, operating system and
screen size of the client device.

We can obtain the browser and operating system by parsing the PHP
variable $_SERVER['HTTP_USER_AGENT']. But we still need the screen
size in order to correct the incompatible items correctly.

To do this, we use JavaScript in the client to save the JavaScript
variables screen.width and screen.height as cookies. Then we use
JavaScript to reload the page. This provides the necessary persistence
of the JavaScript variables so that PHP can obtain their values by
reading the appropriate cookies.

Now that PHP can obtain the required information, we modify our CSS
based on the values of the PHP variables. We define PHP variables
for the sizes of the various on-screen elements using the desired
default values. Then we use a PHP switch statement to override the
default values based on the browser, operating system and screen size.

It might be argued that it would be impossible to handle all the
possible screen sizes. That's probably true. Would we actually need to
cover every possible screen size? Probably not, since older devices will
be retired when the user switches to a newer device. If we limit our code
to the top 10 or 15 most recent devices, the task becomes much more
manageable. A switch statement with 10 to 15 cases should be able to
cover the majority of devices currently in use.

Currently, we are using a switch statement with 11 cases. As newer
devices appear, the switch can be extended to cover those devices.

//  define the PHP variables based on the default screen size
//  override the variables within the switch case required by the
//  device in use

switch($screenWidth) {
      case 360:      //  Android
         //     case 1
         break;
      case 375:      //  iPad, iPhone
         //     case 2
         break;
      case 390:      //  iPad, iPhone
         //     case 3
         break;
      case 414:      //  iPad, iPhone
         //     case 4
        break;
      case 768:      //  iPad
         //     case 5
         break;
      case 820:      //  iPad
         //     case 6
         break;
      case 1133:     //  Android
         //     case 7
         break;
      case 1143:     //  Android
         //     case 8
         break;
      case 1366:     //  Win10
         //     case 9
         break;
      case 1536:     //  Win10
         //     case 10
         break;
      case 1707:     //  Win10
         //     case 11
         break;
      default:

   //  end switch
}