roll back to previous version
Github Actions is failing because it says the download and upload artifacts
is v2. I have edited the workflow file to set it at v4 and it is
still failing.

Yet when I revert the files to the previous version, whi;e keeping v4 in
the workflow file, everything is building correctly.